requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-multipart==0.0.18
sqlalchemy==2.0.36
aiosqlite==0.20.0
anthropic==0.42.0
pillow==11.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.4
pydantic-settings==2.7.1
aiofiles==24.1.0
